🚚 move utils ordner and update import paths
